refactor(Table): remove the interface to index file
