--- src/components/data/Table/interfaces/Table.Action.interface.ts
@@ -1,5 +1,0 @@
-export interface IAction {
-  id: string;
-  actionName: string;
-  content: (entry: any) => JSX.Element;
-}
