--- src/components/data/Table/interfaces/Table.Breakpoint.interface.ts
@@ -1,4 +1,0 @@
-export interface IBreakpoint {
-  breakpoint: string;
-  totalColumns: number;
-}
