--- src/components/data/Table/interfaces/Table.DisplayEntry.interface.ts
@@ -1,13 +1,0 @@
-import { IAction } from "./Table.Action.interface";
-import { IEntry } from "./Table.Entry.interface";
-import { ITitle } from "./Table.Title.interface";
-
-export interface IDisplayEntryProps {
-  portalId: string;
-  entry: IEntry;
-  actions: IAction[];
-  title: string;
-  titleLabels: ITitle[];
-  infoTitle?: string;
-  actionsTitle?: string;
-}
