--- src/components/data/Table/interfaces/Table.Entry.interface.ts
@@ -1,4 +1,0 @@
-export interface IEntry {
-  id: string;
-  [key: string]: any;
-}
