--- src/components/data/Table/interfaces/Table.Pagination.interface.ts
@@ -1,9 +1,0 @@
-export interface IPaginationProps {
-  firstEntryInPage: number;
-  lastEntryInPage: number;
-  totalRecords: number;
-  handleStartPage: () => void;
-  handlePrevPage: () => void;
-  handleNextPage: () => void;
-  handleEndPage: () => void;
-}
