--- src/components/data/Table/interfaces/Table.Title.interface.ts
@@ -1,5 +1,0 @@
-export interface ITitle {
-  id: string;
-  titleName: string;
-  priority: number;
-}
